Improve txDetails code
