+++ app/connectors/index.js
@@ -28,7 +28,6 @@ export { default as ticketsOverview } from "./ticketsOverview";
 export { default as ticketsPage } from "./ticketsPage";
 export { default as transactionDetails } from "./transactionDetails";
 export { default as transactionLink } from "./transactionLink";
-export { default as transactionPage } from "./transactionPage";
 export { default as validateAddressPage } from "./validateAddressPage";
 export { default as verifyMessagePage } from "./verifyMessagePage";
 export { default as votingPrefs } from "./votingPrefs";
