+++ app/containers/Wallet.js
@@ -62,7 +62,7 @@ class Wallet extends React.Component {
             <Route path="/trezor"                         component={TrezorPage} />
             <Route path="/ln"                             component={LNPage} />
           </MainSwitch>
-          <Route path="/transactions/history/:txHash" component={TransactionPage} />
+          <Route path="/transaction/history/:txHash" component={TransactionPage} />
           <Route path="/proposal/details/:token" component={ProposalDetailsPage}/>
         </BlurableContainer>
       </div>
