+++ app/index.js
@@ -217,10 +217,6 @@ var initialState = {
     // values are arrays of account numbers.
     maturingBlockHeights: {},
 
-    // list of outstanding requests for additional stake data from transactions
-    // (indexed by transaction hash)
-    fetchMissingStakeTxDataAttempt: {},
-
     // Shown under governance tab
     treasuryBalance: null
   },
